gliomaSet(rbsurv)	R Documentation

Gene expression and survival data of the patients with gliomas

Description

These data sets consist of gene expression and survival of the patients with gliomas. Note that it contains a subset of the data published in Freije et al. (2004).

Source

Freije et al. (2004). Gene Expression Profiling of Gliomas Strongly Predicts Survival, Cancer Research, 64: 6503-6510.

[Package rbsurv version 2.0.0 Index]