rbsurv(rbsurv)	R Documentation

Robust likelihood-based survival modeling

Description

This selects survival-associated genes with microarray data.

Usage


        rbsurv(time, ...)

Arguments

time 	 an object for which the extraction of model rbsurv is meaningful.

... 	 other arguments

Author(s)

HyungJun Cho, Sukwoo Kim, Soo-heang Eo, and Jaewoo Kang

References

Cho,H., Yu,A., Kim,S., Kang,J., and Hong S-M. (2009). Robust likelihood-based survival modeling for microarray gene expression Data, Journal of Statistical Software, 29(1):1-16. URL http://www.jstatsoft.org/v29/i01/.

See Also

rbsurv.default

Examples



library(rbsurv)
data(gliomaSet)
x <- exprs(gliomaSet)
x <- log2(x)
time <- gliomaSet$Time
status <- gliomaSet$Status
z <- cbind(gliomaSet$Age, gliomaSet$Gender) 

fit <- rbsurv(time=time, status=status, x=x,  method="efron", max.n.genes=20, n.iter=10, n.fold=3,  n.seq=1)
fit$model


[Package rbsurv version 2.0.0 Index]