rbsurv.default(rbsurv)	R Documentation

Robust likelihood-based survival modeling

Description

This selects survival-associated genes with microarray data.

Usage


    ## Default S3 method:
    rbsurv(time, status, x, z=NULL, alpha=1, gene.ID=NULL, method="efron",
                   n.iter=10, n.fold=3,  n.seq=1,  seed=1234, max.n.genes=nrow(x),...)

Arguments

time 	 a vector for survival times

status 	 a vector for survival status, 0=censored, 1=event

x 	 a matrix for expression values (genes in rows, samples in columns)

z 	 a matrix for risk factors

alpha 	 significance level for evaluating risk factors; significant risk factors included with the alpha level if alpha < 1

gene.ID 	 a vector for gene IDs; if NULL, row numbers are assigned.

method 	 a character string specifying the method for tie handling. Choose one of "efron", "breslow", "exact". The default is "efron". If there are no tied death times all the methods are equivalent.

n.iter 	 the number of iterations for gene selection

n.fold 	 the number of partitions of samples

n.seq 	 the number of sequential runs or multiple models

seed 	 a seed for sample partitioning

max.n.genes 	 the maximum number of genes considered. If the number of the input genes is greater than the given number, it is reduced by fitting individual Cox models.

... 	 other arguments

Value

model 	 survival-associated gene model

n.genes 	 number of genes

n.samples 	 number of samples

method 	 method for tie handling

covariates 	 covariates

n.iter 	 number of iterations for gene seletion

n.fold 	 number of partitions of samples

n.seq 	 number of sequential runs or multiple models

gene.list 	 a list of genes included in the models

Author(s)

HyungJun Cho, Sukwoo Kim, Soo-heang Eo, and Jaewoo Kang

References

Cho,H., Yu,A., Kim,S., Kang,J., and Hong S-M. (2009). Robust likelihood-based survival modeling for microarray gene expression Data, Journal of Statistical Software, 29(1):1-16. URL http://www.jstatsoft.org/v29/i01/.

See Also

rbsurv

[Package rbsurv version 2.0.0 Index]